Implemented custom exceptions and responses and concluded the application project

+++ HR.LeaveManagement/HR.LeaveManagement/src/Core/HR.LeaveManagement.Application/Features/LeaveAllocations/Handlers/Commands/DeleteLeaveAllocationCommandHandler.cs
@@ -1,6 +1,8 @@
 using AutoMapper;
+using HR.LeaveManagement.Core.HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Core.HR.LeaveManagement.Application.Features.LeaveAllocations.Requests.Commands;
 using HR.LeaveManagement.Core.HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Core.HR.LeaveManagement.Domain;
 using MediatR;
 
 namespace HR.LeaveManagement.Core.HR.LeaveManagement.Application.Features.LeaveAllocations.Handlers.Commands;
@@ -19,6 +21,9 @@ public class DeleteLeaveAllocationCommandHandler : IRequestHandler<DeleteLeaveAl
     public async Task<Unit> Handle(DeleteLeaveAllocationCommand request, CancellationToken cancellationToken)
     {
         var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+
+        if (leaveAllocation == null)
+            throw new NotFoundException(nameof(LeaveAllocation), request.Id);
         
         await _leaveAllocationRepository.Delete(leaveAllocation);
