+++ HR.LeaveManagement/HR.LeaveManagement/src/Core/HR.LeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,6 +1,8 @@
 using AutoMapper;
+using HR.LeaveManagement.Core.HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Core.HR.LeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HR.LeaveManagement.Core.HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Core.HR.LeaveManagement.Domain;
 using MediatR;
 
 namespace HR.LeaveManagement.Core.HR.LeaveManagement.Application.Features.LeaveRequests.Handlers.Commands;
@@ -19,6 +21,9 @@ public class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveReque
     public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
     {
         var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+        
+        if (leaveRequest == null)
+            throw new NotFoundException(nameof(LeaveRequest), request.Id);
         
         await _leaveRequestRepository.Delete(leaveRequest);
