+++ HR.LeaveManagement/HR.LeaveManagement/src/Core/HR.LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,6 +1,8 @@
 using AutoMapper;
+using HR.LeaveManagement.Core.HR.LeaveManagement.Application.Exceptions;
 using HR.LeaveManagement.Core.HR.LeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
 using HR.LeaveManagement.Core.HR.LeaveManagement.Application.Persistence.Contracts;
+using HR.LeaveManagement.Core.HR.LeaveManagement.Domain;
 using MediatR;
 
 namespace HR.LeaveManagement.Core.HR.LeaveManagement.Application.Features.LeaveTypes.Handlers.Commands;
@@ -19,6 +21,9 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
         var leaveType = await _leaveTypeRepository.Get(request.Id);
+        
+        if (leaveType == null)
+            throw new NotFoundException(nameof(LeaveType), request.Id);
         
         await _leaveTypeRepository.Delete(leaveType);
